shader interface update 3

+++ modules/core/include/core/impl/types/memory.hxx
@@ -1,21 +1,21 @@
 #pragma once
 
-
-
+#include <cstdint>
 namespace mlss
 {
     // using booltype = bool;  // Replaced with bool
 
     //=================================================================================================================
-    //                                   ObjectManager                                                                   
+    //                                   ObjectManager
     //=================================================================================================================
-    template<class ObjectType>
+    template <class ObjectType>
     class ObjectManager
     {
     public:
+
         using object_type = ObjectType;
         using pointer = object_type*;
-        using handle_type = std::uint64_t;  // For storing addresses  
+        using handle_type = std::uintptr_t; // For storing addresses
 
         //---------------------------------------------------------------------
         ~ObjectManager() = default;
@@ -27,15 +27,15 @@ namespace mlss
         static handle_type addObject(std::unique_ptr<object_type>&& object);
 
         //---------------------------------------------------------------------
-        template<typename T = object_type>
+        template <typename T = object_type>
         static T* getPointer(handle_type handle);
 
         //---------------------------------------------------------------------
-        template<typename PtrType>
+        template <typename PtrType>
         static bool isInitialized(PtrType* ptr);
 
         //---------------------------------------------------------------------
-        template<typename PtrType>
+        template <typename PtrType>
         static void markAsInitialized(PtrType* ptr);
 
     private:
@@ -44,20 +44,20 @@ namespace mlss
         static ObjectManager& createInstance();
 
         //---------------------------------------------------------------------
-        template<typename PtrType>
+        template <typename PtrType>
         bool isInitialized_(PtrType* ptr) const;
 
         //---------------------------------------------------------------------
-        template<typename PtrType>
+        template <typename PtrType>
         void markAsInitialized_(PtrType* ptr);
 
         //---------------------------------------------------------------------
         std::unordered_set<std::unique_ptr<object_type>> m_objects;
-        std::unordered_set<const void*> m_initialized_ptr;  // Store void* for any pointer level  
+        std::unordered_set<const void*> m_initialized_ptr; // Store void* for any pointer level
         std::mutex m_mutex;
     };
 
     //=================================================================================================================
     using MemoryManager = ObjectManager<Any>;
 
-} // mlss
+} // namespace mlss
